Camera : allow camera to use power key as shutter
Provides a way for an app to take control of the power key.
Used by the camera to make the power key control the shutter.

Change-Id: I85a1e1761199f4604672be42a3a5005227f5451a
(cherry picked from commit 15661444ae1faea831218f0c936b756de2f0698b)

Conflicts:
	core/java/android/view/WindowManager.java
	core/res/res/values/cm_strings.xml

Fix typo in @hide annotation for PRIVATE_FLAG_PREVENT_POWER_KEY

Change-Id: I19622ffe177553cd5a80022e76a54f8e16e49cac

PhoneWindowPolicy: Fix NPE race condition by excessive screen on delay

When screen on has an excessive delay mTopFullscreenOpaqueWindowState could
not be initialized. This derives in a NPE from a Power key press. If
mTopFullscreenOpaqueWindowState = null then PREVENT_POWER_KEY can be safelly
ignored

Change-Id: I04ce5d88f6f9abf02b2bbcdce4a2acad2f6451fa
Signed-off-by: Jorge Ruesga <jorge@ruesga.com>

Prevent power key capture when screen is off

The ability for an activity to capture the power key, which was
added to support power key as shutter in the camera, should only
allow the capture when the screen is on. Otherwise, if an activity
that captures the power key is to the front when the device turns
off, the user will be unable to turn it back on.

Change-Id: Ib119d6914ec72554b404c1cc17eef3a932d5d402