Telephony: Fix signal annunciators missing issue.
When sim detection is delayed during power up, default subid
is first set to dummy subid. Later when sim is detected, default
sub id is set to the correct value. Due to a wrong check for
old and new subids, the correct value is not set in
TelephonyRegistry. This causes callbacks to be not
invoked for listeners with default sub id.

Change-Id: I00da4de5323125f7aefebd6fa1fb8f787e43f8f9
CRs-Fixed: 809783